refactor: type 분리
